split integration and functional tests
